Move CustomComboRow to Widgets directory
